+++ haskell/app/Trader/Metrics.hs
@@ -125,14 +125,19 @@ computeMetrics periodsPerYear br =
 
 returnsFromEquity :: [Double] -> [Double]
 returnsFromEquity eq =
-  case eq of
+  case eq' of
     [] -> []
     [_] -> []
-    _ -> zipWith ret eq (tail eq)
+    _ -> zipWith ret eq' (tail eq')
   where
     bad x = isNaN x || isInfinite x
+    clamp x =
+      if bad x || x < 0
+        then 0
+        else x
+    eq' = map clamp eq
     ret a b =
-      if bad a || bad b || a <= 0
+      if a <= 0
         then 0
         else b / a - 1